qpsmtpd: fix accepting all for relay_clients
When a user is authenticated before this change they could send
absolutely anything at qpsmtpd which would then have it refused by
postfix giving a not overly useful 451 Can't queue message error.
Change the behaviour so the skip for relay clients doesn't happen
until the very end.

This change also introduces a new return code for validrcptto, which
is given when the domain is hosted but the user doesn't exist.