tests: add some improvements to the base classes
Add our own assertRaises method which allows a test to check that a call
triggers a specific exception (or a number of exceptions), this will
also silence gevent from printing messages to stderr.

On top of this also modify the way threaded database access is managed,
using the sharing between threads since we know that it will only be
being accessed by one thread at one time.