queue: log messages sent by auth'd users
Add in a log of the messages sent by users that have authenticated
with the smtp server. This will allow for a rate limit check to be
put in place to prevent malicious users (either through their account
being compromised or registering with intent) to be stopped quickly
and the appropriate action taken.