core: send out an alert when limit hit
To allow a malicious user or a user who's account has been compromised
send out an alert each time the sending limit is hit. This will most
likely have to be improved to send it out at most up to once per hour,
as well as including automatic account disabling.